refactor: auth 패키지 생성 및 책임 분리
