+++ src/main/java/roomescape/member/service/MemberService.java
@@ -1,82 +1,23 @@
 package roomescape.member.service;
 
-import jakarta.servlet.http.Cookie;
 import org.springframework.stereotype.Service;
-import roomescape.exception.BadRequestException;
-import roomescape.exception.ExceptionMessage;
 import roomescape.member.dao.MemberDao;
 import roomescape.member.domain.Member;
 import roomescape.member.domain.Role;
-import roomescape.member.dto.request.LoginRequest;
 import roomescape.member.dto.request.MemberRequest;
-import roomescape.member.dto.response.LoginCheckResponse;
-import roomescape.member.dto.response.LoginResponse;
 import roomescape.member.dto.response.MemberResponse;
-import roomescape.member.util.JwtTokenProvider;
 
 @Service
 public class MemberService {
 
-    private final JwtTokenProvider jwtTokenProvider;
     private final MemberDao memberDao;
 
-    public MemberService(JwtTokenProvider jwtTokenProvider, MemberDao memberDao) {
-        this.jwtTokenProvider = jwtTokenProvider;
+    public MemberService(MemberDao memberDao) {
         this.memberDao = memberDao;
     }
 
     public MemberResponse createMember(MemberRequest memberRequest) {
         Member member = memberDao.save(new Member(memberRequest.getName(), memberRequest.getEmail(), memberRequest.getPassword(), Role.USER));
         return new MemberResponse(member.getId(), member.getName(), member.getEmail());
-    }
-
-    public LoginResponse login(LoginRequest request) {
-        validateLoginValues(request);
-        Member member = getMemberWithLogin(request);
-        String accessToken = jwtTokenProvider.createAccessToken(member);
-        return new LoginResponse(accessToken);
-    }
-
-    private void validateLoginValues(LoginRequest request) {
-        validateEmailNotBlank(request.email());
-        validatePasswordNotBlank(request.password());
-    }
-
-    private void validateEmailNotBlank(String email) {
-        if (email == null || email.isBlank()) {
-            throw new BadRequestException(ExceptionMessage.INVALID_EMAIL.getMessage());
-        }
-    }
-
-    private void validatePasswordNotBlank(String password) {
-        if (password == null || password.isBlank()) {
-            throw new BadRequestException(ExceptionMessage.INVALID_PASSWORD.getMessage());
-        }
-    }
-
-    private Member getMemberWithLogin(LoginRequest request) {
-        return memberDao.findByEmailAndPassword(request.email(), request.password())
-                .orElseThrow(() -> new BadRequestException(ExceptionMessage.MEMBER_NOT_FOUND.getMessage()));
-    }
-
-    public LoginCheckResponse loginCheck(Cookie cookie) {
-        validateCookie(cookie);
-        String accessToken = cookie.getValue();
-        long memberId = jwtTokenProvider.parseToken(accessToken);
-        Member member = getMemberById(memberId);
-        return new LoginCheckResponse(member);
-    }
-
-    private void validateCookie(Cookie cookie) {
-        if (cookie == null
-                || cookie.getValue() == null
-                || cookie.getValue().isBlank()) {
-            throw new BadRequestException(ExceptionMessage.COOKIE_NOT_FOUND.getMessage());
-        }
-    }
-
-    private Member getMemberById(long memberId) {
-        return memberDao.findById(memberId)
-                .orElseThrow(() -> new BadRequestException(ExceptionMessage.MEMBER_NOT_FOUND.getMessage()));
     }
 }
